Rethink the whole idea of how this should(tm) work
