+++ Live.Avalonia.Sample/Live.Avalonia.Sample.csproj
@@ -16,9 +16,9 @@
         <Watch Include="**\*.xaml" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Avalonia" Version="0.9.0" />
-        <PackageReference Include="Avalonia.Desktop" Version="0.9.0" />
-        <PackageReference Include="Avalonia.ReactiveUI" Version="0.9.0" />
+        <PackageReference Include="Avalonia" Version="0.9.9" />
+        <PackageReference Include="Avalonia.Desktop" Version="0.9.9" />
+        <PackageReference Include="Avalonia.ReactiveUI" Version="0.9.9" />
         <PackageReference Include="Citrus.Avalonia" Version="1.2.6" />
     </ItemGroup>
     <ItemGroup>
