+++ Live.Avalonia/ILiveView.cs
@@ -1,0 +1,9 @@
+using Avalonia.Controls;
+
+namespace Live.Avalonia
+{
+    public interface ILiveView
+    {
+        object CreateView(Window window);
+    }
+}
